fixed delete account issues

+++ delete-account.php
@@ -6,16 +6,23 @@ session_start();
 
 $user_name = $_SESSION['user_name'];
 
-$sql = "DELETE FROM `answers` WHERE `users_user_name` = '{$user_name}';";
-run_sql($conn, $sql);
+if ($user_name != 'ccatura') {
+    $sql = "DELETE FROM `answers` WHERE `users_user_name` = '{$user_name}';";
+    run_sql($conn, $sql);
 
-$sql = "DELETE FROM `users` WHERE `user_name` = '{$user_name}';";
-run_sql($conn, $sql);
+    $sql = "DELETE FROM `messages` WHERE `user_name_to` = '{$user_name}';";
+    run_sql($conn, $sql);
 
-$_SESSION['message'] = "<div>Account <strong>{$user_name}</strong> deleted . Sorry to see you go, dude.</div>";
-echo $sql . '<br>';
-echo $_SESSION['message'];
+    $sql = "DELETE FROM `messages` WHERE `user_name_from` = '{$user_name}';";
+    run_sql($conn, $sql);
 
+    $sql = "DELETE FROM `users` WHERE `user_name` = '{$user_name}';";
+    run_sql($conn, $sql);
+
+    $_SESSION['message'] = "<div>Account <strong>{$user_name}</strong> deleted . Sorry to see you go, dude.</div>";
+    echo $sql . '<br>';
+    echo $_SESSION['message'];
+}
 
 header("Location: ./?session=false");
 // header("Location: ./");
